Fix autoforward
The old GetMovementInputVector function that gets the input in local space is not used anymore by the game.
Now it uses the one that gets the input in world space. This sucks for some of our features.
To fix this, hook one step deeper into the individual GetInputValue calls.
They basically return a float, yet they return a float Vector where x=y.